two ways to create tables in tests why
