+++ posthog/test/base.py
@@ -2,6 +2,7 @@ import datetime as dt
 import inspect
 import re
 import resource
+import sys
 import threading
 import time
 import uuid
@@ -34,17 +35,13 @@ from posthog.cloud_utils import (
 )
 from posthog.models import Dashboard, DashboardTile, Insight, Organization, Team, User
 from posthog.models.channel_type.sql import (
-    CHANNEL_DEFINITION_TABLE_SQL,
     DROP_CHANNEL_DEFINITION_TABLE_SQL,
     DROP_CHANNEL_DEFINITION_DICTIONARY_SQL,
-    CHANNEL_DEFINITION_DICTIONARY_SQL,
-    CHANNEL_DEFINITION_DATA_SQL,
 )
 from posthog.models.cohort.sql import TRUNCATE_COHORTPEOPLE_TABLE_SQL
 from posthog.models.event.sql import (
     DISTRIBUTED_EVENTS_TABLE_SQL,
     DROP_EVENTS_TABLE_SQL,
-    EVENTS_TABLE_SQL,
 )
 from posthog.models.event.util import bulk_create_events
 from posthog.models.group.sql import TRUNCATE_GROUPS_TABLE_SQL
@@ -53,7 +50,6 @@ from posthog.models.organization import OrganizationMembership
 from posthog.models.person import Person
 from posthog.models.person.sql import (
     DROP_PERSON_TABLE_SQL,
-    PERSONS_TABLE_SQL,
     TRUNCATE_PERSON_DISTINCT_ID2_TABLE_SQL,
     TRUNCATE_PERSON_DISTINCT_ID_TABLE_SQL,
     TRUNCATE_PERSON_STATIC_COHORT_TABLE_SQL,
@@ -64,20 +60,16 @@ from posthog.models.sessions.sql import (
     DROP_SESSION_TABLE_SQL,
     DROP_SESSION_MATERIALIZED_VIEW_SQL,
     DROP_SESSION_VIEW_SQL,
-    SESSIONS_TABLE_SQL,
     SESSIONS_TABLE_MV_SQL,
-    SESSIONS_VIEW_SQL,
     DISTRIBUTED_SESSIONS_TABLE_SQL,
 )
 from posthog.session_recordings.sql.session_recording_event_sql import (
     DISTRIBUTED_SESSION_RECORDING_EVENTS_TABLE_SQL,
     DROP_SESSION_RECORDING_EVENTS_TABLE_SQL,
-    SESSION_RECORDING_EVENTS_TABLE_SQL,
 )
 from posthog.session_recordings.sql.session_replay_event_sql import (
     DISTRIBUTED_SESSION_REPLAY_EVENTS_TABLE_SQL,
     DROP_SESSION_REPLAY_EVENTS_TABLE_SQL,
-    SESSION_REPLAY_EVENTS_TABLE_SQL,
 )
 from posthog.settings.utils import get_from_env, str_to_bool
 from posthog.test.assert_faster_than import assert_faster_than
@@ -886,47 +878,53 @@ class ClickhouseDestroyTablesMixin(BaseTest):
     """
 
     def _table_operations(self):
+        # truncating is fast(er)
         run_clickhouse_statement_in_parallel(
             [
-                DROP_EVENTS_TABLE_SQL(),
-                DROP_PERSON_TABLE_SQL,
-                TRUNCATE_PERSON_DISTINCT_ID_TABLE_SQL,
-                TRUNCATE_PERSON_DISTINCT_ID2_TABLE_SQL,
-                DROP_SESSION_RECORDING_EVENTS_TABLE_SQL(),
-                DROP_SESSION_REPLAY_EVENTS_TABLE_SQL(),
                 TRUNCATE_GROUPS_TABLE_SQL,
                 TRUNCATE_COHORTPEOPLE_TABLE_SQL,
                 TRUNCATE_PERSON_STATIC_COHORT_TABLE_SQL,
                 TRUNCATE_PLUGIN_LOG_ENTRIES_TABLE_SQL,
-                DROP_CHANNEL_DEFINITION_TABLE_SQL,
-                DROP_CHANNEL_DEFINITION_DICTIONARY_SQL,
-                DROP_SESSION_TABLE_SQL(),
-                DROP_SESSION_MATERIALIZED_VIEW_SQL(),
-                DROP_SESSION_VIEW_SQL(),
+                TRUNCATE_PERSON_DISTINCT_ID_TABLE_SQL,
+                TRUNCATE_PERSON_DISTINCT_ID2_TABLE_SQL,
             ]
         )
+
+        # drop materialized views first
         run_clickhouse_statement_in_parallel(
             [
-                EVENTS_TABLE_SQL(),
-                PERSONS_TABLE_SQL(),
-                SESSION_RECORDING_EVENTS_TABLE_SQL(),
-                SESSION_REPLAY_EVENTS_TABLE_SQL(),
-                CHANNEL_DEFINITION_TABLE_SQL(),
-                CHANNEL_DEFINITION_DICTIONARY_SQL,
-                SESSIONS_TABLE_SQL(),
+                SESSIONS_TABLE_MV_SQL(),
+                DROP_SESSION_MATERIALIZED_VIEW_SQL(),
             ]
         )
+
+        # then we'd do kafka if we dropped any
+
+        # I guess distributed tables should go before the local tables they front
         run_clickhouse_statement_in_parallel(
             [
                 DISTRIBUTED_EVENTS_TABLE_SQL(),
                 DISTRIBUTED_SESSION_RECORDING_EVENTS_TABLE_SQL(),
                 DISTRIBUTED_SESSION_REPLAY_EVENTS_TABLE_SQL(),
-                CHANNEL_DEFINITION_DATA_SQL,
-                SESSIONS_TABLE_MV_SQL(),
-                SESSIONS_VIEW_SQL(),
                 DISTRIBUTED_SESSIONS_TABLE_SQL(),
             ]
         )
+
+        run_clickhouse_statement_in_parallel(
+            [
+                DROP_EVENTS_TABLE_SQL(),
+                DROP_PERSON_TABLE_SQL,
+                DROP_SESSION_RECORDING_EVENTS_TABLE_SQL(),
+                DROP_SESSION_REPLAY_EVENTS_TABLE_SQL(),
+                DROP_CHANNEL_DEFINITION_TABLE_SQL,
+                DROP_CHANNEL_DEFINITION_DICTIONARY_SQL,
+                DROP_SESSION_TABLE_SQL(),
+                DROP_SESSION_VIEW_SQL(),
+            ]
+        )
+
+        # now recreate the tables
+        create_clickhouse_tables(sys.maxsize)
 
     def setUp(self):
         super().setUp()
@@ -1063,3 +1061,43 @@ def create_person_id_override_by_distinct_id(
         VALUES ({team_id}, '{person_id_from}', '{person_id_to}', {version})
     """
     )
+
+
+def create_clickhouse_tables(num_tables: int):
+    # Create clickhouse tables to default before running test
+    # Mostly so that test runs locally work correctly
+    from posthog.clickhouse.schema import (
+        CREATE_DISTRIBUTED_TABLE_QUERIES,
+        CREATE_MERGETREE_TABLE_QUERIES,
+        CREATE_MV_TABLE_QUERIES,
+        CREATE_DATA_QUERIES,
+        CREATE_DICTIONARY_QUERIES,
+        CREATE_VIEW_QUERIES,
+        build_query,
+        CREATE_KAFKA_TABLE_QUERIES,
+    )
+
+    # REMEMBER TO ADD ANY NEW CLICKHOUSE TABLES TO THIS ARRAY!
+    CREATE_TABLE_QUERIES: Tuple[Any, ...] = CREATE_MERGETREE_TABLE_QUERIES + CREATE_DISTRIBUTED_TABLE_QUERIES
+
+    # Check if all the tables have already been created
+    if num_tables == len(CREATE_TABLE_QUERIES):
+        return
+
+    table_queries = list(map(build_query, CREATE_TABLE_QUERIES))
+    run_clickhouse_statement_in_parallel(table_queries)
+
+    kafka_queries = list(map(build_query, CREATE_KAFKA_TABLE_QUERIES))
+    run_clickhouse_statement_in_parallel(kafka_queries)
+
+    mv_queries = list(map(build_query, CREATE_MV_TABLE_QUERIES))
+    run_clickhouse_statement_in_parallel(mv_queries)
+
+    view_queries = list(map(build_query, CREATE_VIEW_QUERIES))
+    run_clickhouse_statement_in_parallel(view_queries)
+
+    data_queries = list(map(build_query, CREATE_DATA_QUERIES))
+    run_clickhouse_statement_in_parallel(data_queries)
+
+    dictionary_queries = list(map(build_query, CREATE_DICTIONARY_QUERIES))
+    run_clickhouse_statement_in_parallel(dictionary_queries)
